requirements.txt
aiofiles==24.1.0
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fasttext==0.9.3
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
huggingface-hub==0.29.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
langdetect==1.0.9
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.50.0
priority==2.0.0
pybind11==2.13.6
pyee==12.1.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
pytz==2025.1
PyYAML==6.0.2
Quart==0.20.0
quart-cors==0.8.0
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
six==1.17.0
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tldextract==5.1.3
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wsproto==1.2.0
